New Version with Different Scale
I decided that I did not like the scale the individual was standing on, so with some further cutting and inverting, I was able to place her on a new scale.  I like the idea of using a digital looking font better if I am going to be using words on the scale because this would allow the words to change in different versions of the poster, which is my vision for the project.  I am having trouble with downloading a font that I like, which is my next challenge.  I also need to make a decision of whether I like marble or wood flooring best to contrast the scale, and how I would like to arrange the page with the text "Redefine the scale".  Should I add "one number does not define you" or is a shorter tagline more powerful.