First Attempt at Cropping
In order to change the scale into something that does not read numbers, but rather adjectives, I needed to crop out the center of the scale.  I also cropped out the flooring because I believe different tile would capture the audience better, as well as bring attention to the foreground.