FINAL Project 1
A lot of changes happened between Draft 4 and this final draft.  I finally came to a decision on flooring and decided to add the NEDA Awareness sticker at the bottom right (a recreation of the original because the one on the website was copyrighted and did not allow for reuse).